AuthorizationModificationDone
