+++ Client/src/components/post-ad/PostAd.jsx
@@ -1,10 +1,13 @@
 /* eslint-disable react/no-unescaped-entities */
 /* eslint-disable no-unused-vars */
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Typography } from "@mui/material";
 import CarInfoModal from "../Modal/CarInfoModal";
 import UploadModal from "../Modal/uploadModal";
 import ContactModal from "../Modal/ContactModal";
+import { useNavigate } from "react-router-dom";
+import Cookies from "js-cookie";
+
 import {
   Wrapper,
   SubHeading,
@@ -21,6 +24,15 @@ import ContactPhoneIcon from "@mui/icons-material/ContactPhone";
 import PublishedAdLayout from "./PublishedAdLayout";
 
 const PostAd = () => {
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const authToken = Cookies.get("authToken");
+    if (!authToken) {
+      navigate("/login");
+    }
+  }, [navigate]);
+
   const [modalOpen, setModalOpen] = useState(false);
   const [activeCard, setActiveCard] = useState(null);
   const [cardStatus, setCardStatus] = useState({
